// src/modules/Canvas/enums.ts
export enum BranchMemberTypeEnum {
  Member = "member",
  Role = "role",
}

export enum BranchAccessEnum {
  EDIT = "edit",
  COMMENT = "comment",
  VIEW = "view",
  PRIVATE = "private",
}

export enum MergeRequestStatusEnum {
  OPEN = "OPEN",
  ACCEPTED = "ACCEPTED",
  REJECTED = "REJECTED",
  PARTIALLY_ACCEPTED = "PARTIALLY_ACCEPTED",
}